#314 extracted common fields to abstract class for integration tests
