+++ quolance-api/src/test/java/com/quolance/quolance_api/integration/BaseIntegrationTest.java
@@ -1,0 +1,31 @@
+package com.quolance.quolance_api.integration;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.quolance.quolance_api.helpers.integration.SessionCreationHelper;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.mock.web.MockHttpSession;
+import org.springframework.test.web.servlet.MockMvc;
+import org.testcontainers.junit.jupiter.Testcontainers;
+
+@SpringBootTest
+@AutoConfigureMockMvc
+@Testcontainers
+public abstract class BaseIntegrationTest extends AbstractTestcontainers {
+    @Autowired
+    protected MockMvc mockMvc;
+
+    @Autowired
+    protected ObjectMapper objectMapper;
+
+    protected SessionCreationHelper sessionCreationHelper;
+
+    protected MockHttpSession session;
+
+    @BeforeEach
+    void setupSessionCreationHelper() {
+        sessionCreationHelper = new SessionCreationHelper(mockMvc, objectMapper);
+    }
+}
