+++ quolance-api/src/test/java/com/quolance/quolance_api/integration/tests/UsersControllerIntegrationTest.java
@@ -1,25 +1,19 @@
-package com.quolance.quolance_api.integration;
+package com.quolance.quolance_api.integration.tests;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.quolance.quolance_api.dtos.users.CreateAdminRequestDto;
 import com.quolance.quolance_api.dtos.users.CreateUserRequestDto;
 import com.quolance.quolance_api.dtos.users.UpdateUserPasswordRequestDto;
 import com.quolance.quolance_api.dtos.users.UpdateUserRequestDto;
 import com.quolance.quolance_api.entities.User;
 import com.quolance.quolance_api.entities.enums.Role;
-import com.quolance.quolance_api.helpers.EntityCreationHelper;
-import com.quolance.quolance_api.helpers.SessionCreationHelper;
+import com.quolance.quolance_api.helpers.integration.EntityCreationHelper;
+import com.quolance.quolance_api.integration.BaseIntegrationTest;
 import com.quolance.quolance_api.repositories.UserRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
-import org.springframework.mock.web.MockHttpSession;
 import org.springframework.security.test.context.support.WithMockUser;
-import org.springframework.test.web.servlet.MockMvc;
-import org.testcontainers.junit.jupiter.Testcontainers;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
@@ -28,26 +22,15 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@SpringBootTest
-@AutoConfigureMockMvc
-@Testcontainers
-class UsersControllerIntegrationTest extends AbstractTestcontainers {
-    @Autowired
-    private MockMvc mockMvc;
-
-    @Autowired
-    private ObjectMapper objectMapper;
+class UsersControllerIntegrationTest extends BaseIntegrationTest {
 
     @Autowired
     private UserRepository userRepository;
-
-    private SessionCreationHelper sessionCreationHelper;
 
     @BeforeEach
     void setUp() {
         userRepository.deleteAll();
         userRepository.save(EntityCreationHelper.createAdmin());
-        sessionCreationHelper = new SessionCreationHelper(mockMvc, objectMapper);
     }
 
     @Test
@@ -123,7 +106,7 @@ class UsersControllerIntegrationTest extends AbstractTestcontainers {
                 .temporaryPassword("Test1234")
                 .passwordConfirmation("Test1234")
                 .build();
-        MockHttpSession session = sessionCreationHelper.getSession("admin@test.com", "Password123!");
+        session = sessionCreationHelper.getSession("admin@test.com", "Password123!");
 
         mockMvc.perform(post("/api/users/admin").session(session)
                         .contentType(MediaType.APPLICATION_JSON)
@@ -166,7 +149,7 @@ class UsersControllerIntegrationTest extends AbstractTestcontainers {
                 .firstName("NEW")
                 .lastName("NEW")
                 .build();
-        MockHttpSession session = sessionCreationHelper.getSession("client@test.com", "Password123!");
+        session = sessionCreationHelper.getSession("client@test.com", "Password123!");
 
         // Act
         String response = mockMvc.perform(put("/api/users").session(session)
@@ -198,7 +181,7 @@ class UsersControllerIntegrationTest extends AbstractTestcontainers {
                 .confirmPassword("NewPassword123!")
                 .build();
 
-        MockHttpSession session = sessionCreationHelper.getSession("client@test.com", "Password123!");
+        session = sessionCreationHelper.getSession("client@test.com", "Password123!");
 
         // Act
         mockMvc.perform(patch("/api/users/password").session(session)
@@ -221,7 +204,7 @@ class UsersControllerIntegrationTest extends AbstractTestcontainers {
                 .confirmPassword("test!")
                 .build();
 
-        MockHttpSession session = sessionCreationHelper.getSession("client@test.com", "Password123!");
+        session = sessionCreationHelper.getSession("client@test.com", "Password123!");
 
         // Act
         mockMvc.perform(patch("/api/users/password").session(session)
@@ -234,18 +217,4 @@ class UsersControllerIntegrationTest extends AbstractTestcontainers {
         assertThat(updatedUser.getPassword()).isEqualTo(createdClient.getPassword());
     }
 
-//    private MockHttpSession getSession(String username, String password) throws Exception {
-//        LoginRequestDto loginRequest = new LoginRequestDto();
-//        loginRequest.setUsername(username);
-//        loginRequest.setPassword(password);
-//
-//        MockHttpSession session = (MockHttpSession) mockMvc.perform(post("/api/auth/login")
-//                        .contentType(MediaType.APPLICATION_JSON)
-//                        .content(objectMapper.writeValueAsString(loginRequest)))
-//                .andExpect(status().isOk())
-//                .andReturn()
-//                .getRequest()
-//                .getSession();
-//        return session;
-//    }
 }
